refactor(web): extract project-intake-banner from task detail
